url수정..

+++ src/main/java/com/sp/team32ndproject/match/service/MatchBoardPhotoInfoService.java
@@ -24,15 +24,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Slf4j
 public class MatchBoardPhotoInfoService {
-	 @Autowired
-	    private AmazonS3 amazonS3;
+	@Autowired
+	private AmazonS3 amazonS3;
 
-	    @Value("${upload.file-path}")
-	    private String uploadFilePath;
+	@Value("${upload.file-path}")
+	private String uploadFilePath;
 
-	    @Value("${bucket.file-Path}")
-	    private String s3FilePath;
-
+	@Value("${bucket.file-Path}")
+	private String s3FilePath;
 
 	private final MatchBoardPhotoInfoMapper matchBoardPhotoInfoMapper;
 
@@ -46,13 +45,13 @@ public class MatchBoardPhotoInfoService {
 				String fileName = UUID.randomUUID() + extName;
 				matchPhoto.setMbpFileName(originName);
 				try {
-					   InputStream inputStream = file.getInputStream();
-		                ObjectMetadata metadata = new ObjectMetadata();
-		                metadata.setContentType(file.getContentType());
-		                metadata.setContentLength(file.getSize());
-		                String s3Key = "upload/" + fileName;
-		                amazonS3.putObject("3nd-team3", s3Key, inputStream, metadata);
-		                matchPhoto.setMbpFilePath("https://3nd-team3.s3.ap-northeast-2.amazonaws.com/" + s3Key);
+					InputStream inputStream = file.getInputStream();
+					ObjectMetadata metadata = new ObjectMetadata();
+					metadata.setContentType(file.getContentType());
+					metadata.setContentLength(file.getSize());
+					String s3Key = "upload/" + fileName;
+					amazonS3.putObject("3nd-team3", s3Key, inputStream, metadata);
+					matchPhoto.setMbpFilePath("https://3nd-team3.s3.ap-northeast-2.amazonaws.com/" + s3Key);
 				} catch (IllegalStateException e) {
 					log.error("error => {}", e.getMessage());
 				} catch (IOException e) {
@@ -69,44 +68,46 @@ public class MatchBoardPhotoInfoService {
 	public int updateGoodsFileInfos(int mbNum, List<MatchBoardPhotoInfoVO> matchPhotos) {
 
 		int result = 0;
-		for (MatchBoardPhotoInfoVO matchPhoto : matchPhotos) {
-			if (matchPhoto.getStatus() == Status.DELETE) {
-				String fileName = matchPhoto.getMbpFilePath();
-				int idx = fileName.lastIndexOf("/") + 1;
-				fileName = fileName.substring(idx);
-				File f = new File("upload/" + fileName);
-				if (f.exists()) {
-					f.delete();
+		if (matchPhotos != null && !matchPhotos.isEmpty()) {
+			for (MatchBoardPhotoInfoVO matchPhoto : matchPhotos) {
+				if (matchPhoto.getStatus() == Status.DELETE) {
+					String fileName = matchPhoto.getMbpFilePath();
+					int idx = fileName.lastIndexOf("/") + 1;
+					fileName = fileName.substring(idx);
+					File f = new File("upload/" + fileName);
+					if (f.exists()) {
+						f.delete();
+					}
+					result += matchBoardPhotoInfoMapper.deleteMatchBoardPhotoInfos(matchPhoto.getMbpNum());
+					continue;
 				}
-				result += matchBoardPhotoInfoMapper.deleteMatchBoardPhotoInfos(matchPhoto.getMbpNum());
-				continue;
-			}
-			MultipartFile file = matchPhoto.getFile();
-			if (file != null) {
-				String originName = file.getOriginalFilename(); // abcd.png
-				String extName = originName.substring(originName.lastIndexOf(".")); // .png
-				String fileName = UUID.randomUUID() + extName;
-				matchPhoto.setMbpFileName(fileName);
-				try {
-				      InputStream inputStream = file.getInputStream();
-		                ObjectMetadata metadata = new ObjectMetadata();
-		                metadata.setContentType(file.getContentType());
-		                metadata.setContentLength(file.getSize());
-		                String s3Key = "upload/" + fileName;
-		                amazonS3.putObject("3nd-team3", s3Key, inputStream, metadata);
-		                matchPhoto.setMbpFilePath("https://3nd-team3.s3.ap-northeast-2.amazonaws.com/" + s3Key);
-				} catch (IllegalStateException e) {
-					log.error("file upload error=>{}", e);
-				} catch (IOException e) {
-					log.error("file upload error=>{}", e);
+				MultipartFile file = matchPhoto.getFile();
+				if (file != null) {
+					String originName = file.getOriginalFilename(); // abcd.png
+					String extName = originName.substring(originName.lastIndexOf(".")); // .png
+					String fileName = UUID.randomUUID() + extName;
+					matchPhoto.setMbpFileName(fileName);
+					try {
+						InputStream inputStream = file.getInputStream();
+						ObjectMetadata metadata = new ObjectMetadata();
+						metadata.setContentType(file.getContentType());
+						metadata.setContentLength(file.getSize());
+						String s3Key = "upload/" + fileName;
+						amazonS3.putObject("3nd-team3", s3Key, inputStream, metadata);
+						matchPhoto.setMbpFilePath("https://3nd-team3.s3.ap-northeast-2.amazonaws.com/" + s3Key);
+					} catch (IllegalStateException e) {
+						log.error("file upload error=>{}", e);
+					} catch (IOException e) {
+						log.error("file upload error=>{}", e);
+					}
 				}
-			}
-			matchPhoto.setMbNum(mbNum);
-			if (file != null) {
-				if (matchPhoto.getStatus() == Status.UPDATE) {
-					result += matchBoardPhotoInfoMapper.updateMatchBoardPhotoInfos(matchPhoto);
-				} else {
-					result += matchBoardPhotoInfoMapper.insertMatchBoardPhotoInfo(matchPhoto);
+				matchPhoto.setMbNum(mbNum);
+				if (file != null) {
+					if (matchPhoto.getStatus() == Status.UPDATE) {
+						result += matchBoardPhotoInfoMapper.updateMatchBoardPhotoInfos(matchPhoto);
+					} else {
+						result += matchBoardPhotoInfoMapper.insertMatchBoardPhotoInfo(matchPhoto);
+					}
 				}
 			}
 		}
